+++ src/libsbmlnetwork_render_helpers.cpp
@@ -3,7 +3,7 @@
 #include "libsbmlnetwork_render.h"
 #include "libsbmlnetwork_sbmldocument_render.h"
 #include "libsbmlnetwork_layout_helpers.h"
-#include "colors/libsbmlnetwork_colors.h"
+#include "features/colors/libsbmlnetwork_colors.h"
 #include "features/styles/libsbmlnetwork_styles.h"
 #include "features/error_log/libsbmlnetwork_error_log.h"
 #include "features/defaults/libsbmlnetwork_defaults_render.h"
